test(react-bloc): 🚨 update coverage threshold
