--- packages/react-bloc/src/lib/components/bloc-listener/index.ts
@@ -1,1 +1,0 @@
-export * from './listener';
