+++ packages/react-bloc/src/lib/components/index.ts
@@ -1,1 +1,1 @@
-export * from './bloc-listener';
+export * from './listener';
