+++ packages/react-bloc/src/lib/hooks/defaults.ts
@@ -1,0 +1,3 @@
+export const defaultSuspendWhen = () => false;
+export const defaultListenWhen = () => true;
+export const defaultErrorWhen = () => false;
