+++ packages/react-bloc/src/lib/hooks/useBlocInstance.ts
@@ -1,0 +1,14 @@
+import { BlocBase, ClassType } from '@jacobtipp/bloc';
+import { useContext } from 'react';
+import { contextMap } from '../provider';
+
+export const useBlocInstance = <Bloc extends ClassType<BlocBase<any>>>(
+  bloc: Bloc
+) => {
+  const context = contextMap.get(bloc);
+  if (!context)
+    throw new Error(`${bloc.name} does not exist in the context map.`);
+
+  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+  return useContext(context.blocContext)! as InstanceType<Bloc>;
+};
